requirements.txt
aiohttp>=3.8.4
beautifulsoup4>=4.11.2
requests>=2.28.2
typing-extensions>=4.5.0
colorama>=0.4.6
tqdm>=4.65.0
rich>=13.3.5
python-dotenv>=1.0.0
pydantic>=1.10.7
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
playwright>=1.30.0
pillow>=9.0.0
selenium>=4.0.0